modules/upword/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0
boto3>=1.34.0
# 文档解析库 (按优先级)
markitdown[docx]>=0.0.1a1
mammoth>=1.6.0
python-docx>=1.1.0
